自媒体管理端：文章发布和关联素材库

+++ yz-leadnews-common/src/main/java/com/leadnews/common/article/ArticleConstants.java
@@ -5,7 +5,7 @@ package com.leadnews.common.article;
  * @date 2023/8/15
  */
 public class ArticleConstants {
-    public static final Short LOADTYPE_LOAD_MORE = 1;
-    public static final Short LOADTYPE_LOAD_NEW = 2;
+    public static final Integer LOADTYPE_LOAD_MORE = 1;
+    public static final Integer LOADTYPE_LOAD_NEW = 2;
     public static final String DEFAULT_TAG = "__all__";
 }
