+++ yz-leadnews-service/article-service/src/main/java/com/leadnews/article/service/ApArticleService.java
@@ -16,7 +16,7 @@ public interface ApArticleService extends IService<ApArticle> {
      * @param dto
      * @return
      */
-    ResponseResult load(Short loadtype, ArticleHomeDTO dto);
+    ResponseResult load(Integer loadtype, ArticleHomeDTO dto);
 
 
 }
